+++ lib/features/workout/presentation/widgets/routine_set_row_widget.dart
@@ -8,8 +8,8 @@ import '../../domain/models/set_template.dart';
 import 'set_type_chip.dart';
 import 'package:flutter_lucide/flutter_lucide.dart';
 import '../../../../util/time_util.dart';
-import '../../../../widgets/common/platform_adaptive_pickers.dart'
-    as adaptive_pickers;
+import '../../../../widgets/common/platform_adaptive_pickers.dart' as adaptive_pickers;
+import '../../../../util/design_constants.dart';
 
 class RoutineSetRowWidget extends StatelessWidget {
   final int setIndex;
@@ -216,7 +216,7 @@ class RoutineSetRowWidget extends StatelessWidget {
                     ? IconButton(
                         icon: const Icon(
                           LucideIcons.trash_2,
-                          color: Colors.redAccent,
+                          color: DesignConstants.brandRedColor,
                         ),
                         onPressed: onRemoveSet,
                       )